move TypeScriptFlow to own dir
