+++ packages/quicktype-core/src/language/TypeScriptFlow/index.ts
@@ -1,0 +1,3 @@
+export { FlowRenderer } from "./FlowRenderer";
+export { TypeScriptRenderer } from "./TypeScriptRenderer";
+export { FlowTargetLanguage, TypeScriptTargetLanguage, tsFlowOptions } from "./language";
